chore(query-cache): split database-free storage tests
Generated-By: PostHog Desktop
Task-Id: fc9b903e-683d-471b-bd5f-2bfa0aac5a50
